+++ AdincoPetrovendorDB/AdincoDB/dbo/Stored Procedures/Procs1/USP_SEL_CNH_FormatoPlanes_GastosInversionesBimestre2023.sql
@@ -954,7 +954,7 @@ INSERT INTO #ResultadoMontos
                                                                         'E'  
                                                                     )  
                                         THEN CAST((TTF.MontoRegistro / TTF.TCD)  
-                                                    * (TTF.RC2122 / (F.MontoConIva / TTF.TCD)) AS FLOAT)  
+                                                    * (TTF.RC2122 / (F.MontoConIva / TTF.TCD)) AS DECIMAL(15, 2))  
                                     ELSE  
                                         0  
                                 END  
@@ -1140,7 +1140,7 @@ INSERT INTO #ResultadoMontos
                                     AND TTF.TipoComprobante IN (  
                                                                     'E'  
                                                                 )  
-                                    THEN CAST(TTF.MontoDolares AS FLOAT)  
+                                    THEN CAST(TTF.MontoDolares AS DECIMAL(15, 2))  
                                 ELSE  
                                     0  
                             END
